Resolve rare eof(::ResponseStream) race condition
When streaming a file via `get_streaming` I have repeatedly encountered
situations in which `eof` returns true, but a subsequent read on the stream
would return more data. While `eof(::ResponseStream)` does check `eof(buffer)`
(to verify that the buffer is empty) as well checking that the stream is done,
because the check is not atomic there seems to be a race condition whereby the
buffer may be empty when it is checked (meaning `eof(buffer)` returns true), but
by the time the stream state is checked (immediately thereafter) the stream has
finished and the buffer has been filled (so `eof(buffer)` would no longer return
true). This results in some cases where `eof(::ResponseStream)` temporarily
returns true even though there is still data to be read (and subsequent calls to
`eof(::ResponseStream)` would return false).

Simply swapping the order in which `eof(::ResponseStream)` makes its
checks resolves the issue.